+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added|Changed|Deprecated|Removed|Fixed|Security
 Nothing so far
 
+## 2.12.2 - 2020-01-31
+### Fixed
+- The `before` and `after` filter helpers will now convert ISO date string when the expected
+  value is a `DateTimeInterface`.
+
 ## 2.12.1 - 2020-01-08
 ### Fixed
 - Linter fixes.